feat: redesigned ingesting datasets table
