+++ sites/geohub/src/components/data-upload/IngestingDatasetHeader.svelte
@@ -1,0 +1,31 @@
+<div class="row columns is-vcentered m-0 is-mobile">
+	<div class="column is-9-mobile">
+		<p class="is-size-6 has-text-centered has-text-weight-bold">Name</p>
+	</div>
+	<div class="column is-2">
+		<p class="is-size-6 has-text-weight-bold">Status</p>
+	</div>
+	<div class="column is-1 hidden-mobile">
+		<p class="is-size-6 has-text-weight-bold">Size</p>
+	</div>
+	<div class="column is-2 hidden-mobile">
+		<p class="is-size-6 has-text-weight-bold">Uploaded at</p>
+	</div>
+	<div class="column is-1 hidden-mobile">
+		<p></p>
+	</div>
+</div>
+
+<style lang="scss">
+	.row {
+		border-top: 1px solid gray;
+		border-bottom: 1px solid gray;
+	}
+
+	.hidden-mobile {
+		display: block;
+		@media (max-width: 48em) {
+			display: none;
+		}
+	}
+</style>
